updated flexpack build info support for poetry
